+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Tap a member in group expense breakdown to see which expenses make up their share.
+
 ## [1.4.0] - 2026-08-03
 
 ### Added